chore: refactor user model
